+++ src/main/java/com/tenco/hobby/service/UserService.java
@@ -76,7 +76,7 @@ public class UserService {
 	 * @return userEntity 응답
 	 */
 	@Transactional
-	public User readInfo(Integer id) {
+	public User readInfo(Long id) {
 
 		User userEntity = userRepository.findById(id);
